Conflicts:

	build/Makefile.common
	src/color.cpp
	src/gui.cpp
	src/loader.cpp

+++ src/color.cpp
@@ -137,7 +137,7 @@ vector<ColorEntry> entries;
 
 void scheme_fill(ColorMapper & in, int scheme){
   if(!color_inited)return;
-  if(scheme < 0 || scheme >= entries.size())return;
+  if(scheme < 0 || (unsigned int)scheme >= entries.size())return;
   in = entries[scheme].map;
 };
 
@@ -188,7 +188,7 @@ void color_init(){
 
 vector<string> color_type(){
   vector<string> out;
-  for(int i = 0; i < entries.size(); i++)
+  for(unsigned int i = 0; i < entries.size(); i++)
     out.push_back(entries[i].name);
   return out;
 };
